Conflicts (solved):
	macosx/Wolf4SDL.xcodeproj/project.pbxproj

+++ id_us_1.cpp
@@ -40,6 +40,9 @@
 //			window
 //
 
+#include <ctype.h>
+#include <stdio.h>
+#include <stdlib.h>
 #include "wl_def.h"
 #include "wl_menu.h"
 
@@ -207,7 +210,7 @@ void US_PrintSigned(int32_t n)
 {
 	char	buffer[32];
 
-	US_Print(ltoa(n,buffer,10));
+	US_Print(ltoa((long int)n,buffer,10));
 }
 
 ///////////////////////////////////////////////////////////////////////////
